+++ modules/mono/editor/Godot.NET.Sdk/Godot.SourceGenerators.Tests/ScriptPropertiesGeneratorTests.cs
@@ -7,7 +7,7 @@ public class ScriptPropertiesGeneratorTests
     [Fact]
     public async void ExportedFields()
     {
-        await CSharpSourceGeneratorVerifier<ScriptPropertiesGenerator>.Verify(
+        await CSharpSourceGeneratorVerifier<ScriptPropertiesGeneratorRunner>.Verify(
             new string[] { "ExportedFields.cs", "MoreExportedFields.cs" },
             new string[] { "ExportedFields_ScriptProperties.generated.cs" }
         );
@@ -16,7 +16,7 @@ public class ScriptPropertiesGeneratorTests
     [Fact]
     public async void ExportedProperties()
     {
-        await CSharpSourceGeneratorVerifier<ScriptPropertiesGenerator>.Verify(
+        await CSharpSourceGeneratorVerifier<ScriptPropertiesGeneratorRunner>.Verify(
             "ExportedProperties.cs",
             "ExportedProperties_ScriptProperties.generated.cs"
         );
@@ -25,7 +25,7 @@ public class ScriptPropertiesGeneratorTests
     [Fact]
     public async void OneWayPropertiesAllReadOnly()
     {
-        await CSharpSourceGeneratorVerifier<ScriptPropertiesGenerator>.Verify(
+        await CSharpSourceGeneratorVerifier<ScriptPropertiesGeneratorRunner>.Verify(
             "AllReadOnly.cs",
             "AllReadOnly_ScriptProperties.generated.cs"
         );
@@ -34,7 +34,7 @@ public class ScriptPropertiesGeneratorTests
     [Fact]
     public async void OneWayPropertiesAllWriteOnly()
     {
-        await CSharpSourceGeneratorVerifier<ScriptPropertiesGenerator>.Verify(
+        await CSharpSourceGeneratorVerifier<ScriptPropertiesGeneratorRunner>.Verify(
             "AllWriteOnly.cs",
             "AllWriteOnly_ScriptProperties.generated.cs"
         );
@@ -43,7 +43,7 @@ public class ScriptPropertiesGeneratorTests
     [Fact]
     public async void OneWayPropertiesMixedReadOnlyWriteOnly()
     {
-        await CSharpSourceGeneratorVerifier<ScriptPropertiesGenerator>.Verify(
+        await CSharpSourceGeneratorVerifier<ScriptPropertiesGeneratorRunner>.Verify(
             "MixedReadOnlyWriteOnly.cs",
             "MixedReadOnlyWriteOnly_ScriptProperties.generated.cs"
         );
@@ -52,7 +52,7 @@ public class ScriptPropertiesGeneratorTests
     [Fact]
     public async void ScriptBoilerplate()
     {
-        await CSharpSourceGeneratorVerifier<ScriptPropertiesGenerator>.Verify(
+        await CSharpSourceGeneratorVerifier<ScriptPropertiesGeneratorRunner>.Verify(
             "ScriptBoilerplate.cs",
             "ScriptBoilerplate_ScriptProperties.generated.cs", "OuterClass.NestedClass_ScriptProperties.generated.cs"
         );
@@ -61,7 +61,7 @@ public class ScriptPropertiesGeneratorTests
     [Fact]
     public async void AbstractGenericNode()
     {
-        await CSharpSourceGeneratorVerifier<ScriptPropertiesGenerator>.Verify(
+        await CSharpSourceGeneratorVerifier<ScriptPropertiesGeneratorRunner>.Verify(
             "AbstractGenericNode.cs",
             "AbstractGenericNode(Of T)_ScriptProperties.generated.cs"
         );
