+++ modules/mono/editor/Godot.NET.Sdk/Godot.SourceGenerators/ScriptMethodsGenerator.cs
@@ -7,7 +7,7 @@ using Microsoft.CodeAnalysis.Text;
 
 namespace Godot.SourceGenerators
 {
-    public class ScriptMethodsGeneratorImplementation : IGeneratorImplementation
+    public class ScriptMethodsGenerator : IGeneratorImplementation
     {
         public void Execute(IGeneratorExecutionContext context)
         {
