+++ modules/mono/editor/Godot.NET.Sdk/Godot.SourceGenerators/ScriptRegistrarGenerator.cs
@@ -4,7 +4,7 @@ namespace Godot.SourceGenerators
 {
     // Placeholder. Once we switch to native extensions this will act as the registrar for all
     // user Godot classes in the assembly. Think of it as something similar to `register_types`.
-    public class ScriptRegistrarGeneratorImplementation //: IGeneratorImplementation
+    public class ScriptRegistrarGenerator //: IGeneratorImplementation
     {
         public void Execute(IGeneratorExecutionContext context)
         {
